Users/ADMIN/Documents/NetBeansProjects/ExamenPIII/src/da/Nodo.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package da;

/**
 *
 * @author ADMIN
 */
public class Nodo {

    private Jugador dato;
    private Nodo sig;
    private Nodo ant;

    public Nodo(Jugador dato) {
        this.dato = dato;
        this.sig = null;
        this.ant = null;
    }

    public Jugador getDato() {
        return dato;
    }

    public void setDato(Jugador dato) {
        this.dato = dato;
    }

    public Nodo getSig() {
        return sig;
    }

    public void setSig(Nodo sig) {
        this.sig = sig;
    }

    public Nodo getAnt() {
        return ant;
    }

    public void setAnt(Nodo ant) {
        this.ant = ant;
    }

}
